fix(templates): move Getting Started into Technical Documentation
Getting Started (install steps, theming, AI assistance card) belongs
in Technical Documentation, not the Overview.

Made-with: Cursor
